Handle 'nesting_level' in PU_hook
This was anyway an oversight, as we seem to only track top-level queries
here, while there are quite a lot of executor commands wrapped by an
utility one.
Originally, however, this commit addresses the problem of the current
multi-level query handling implementation based on a temporary "static
variable" storage. It almost always works, except when an executor query is
wrapped by a cursor's lifetime given cursor command actually has queryId assigned (loops
inside plpg block). So far, I can only think of a loop inside a plpg - simple declare cursor
doesn't have queryId defined. In this case static variable keeps storing the inner
command's queryid, which is stored in the htab entry's payload inside the ExecutorEnd
hook, while the entry's key used for lookup is the proper one, so we
don't find an entry for the outer query each time and store it with the
inner's query queryid as a key.